[enh] Extraction: refactor, using a new ExtractionManager [enh] Extraction: allow to link a product to a parent product [fix] Extraction: fix unit tests [fix] Extraction: fix PGSQL aggregation queries

+++ sumaris-core/src/test/java/net/sumaris/core/dao/DatabaseFixtures.java
@@ -293,14 +293,14 @@ public class DatabaseFixtures {
 
 	/* -- product -- */
 
-	public Integer getProductId(int index) {
+	public Long getProductId(int index) {
 		Preconditions.checkArgument(index >= 0);
 		switch (index) {
 			case 0:
-				return 1;
+				return 1L;
 
 			default:
-				return 1;
+				return 1L;
 		}
 	}
